fix /root paths
The `codeowners` package expects to be passed in paths like:

```
foo/bar.txt
```

not paths like:

```
/foo/bar.txt
```

When you pass in the second kind, I believe it assumes you mean
`/foo/bar.txt` from the root of the whole computer. This means only
rules like `foo/bar.txt` will work, not `/foo/bar.txt` which is looking
from the root of the repo.

In effect this means CODEOWNERS lines starting with `/` are ignored.

To fix this, pass paths like `foo/bar.txt` to codeowners, instead of
`/foo/bar.txt`.